#520 Move some common offset management code to OffsetManagerUtils.
